add cartridge memory

+++ src/main/java/fr/faylixe/jgb/memory/IAddressable.java
@@ -16,6 +16,6 @@ public interface IAddressable {
 	 * 
 	 * @return
 	 */
-	short getOffset();
+	int getOffset();
 
 }
